+++ CHANGES.md
@@ -16,6 +16,7 @@
 * Update urllib3 2.5.0 (98a60f9) to 2.6.2 (83f8643)
 * Update zoneinfo 2025a to 2025c (fallback)
 * Change improve Proper and Series name search for Newznab provider type
+* Fix id parsing in specific cases on newznab providers
 
 
 [develop]